Fix the connection try block and remove the tests associated with it.
I figured out what I was doing wrong when connecting.  I never provided a user object to the session.  Instead of throwing an exception when the hostkey is unknown, jsch will instead use the user object to ask about adding the host key to the host file.  This simplifies things greatly.