Add the recursive directory delete. Remove the code that does not compile added in commit a356a62.
The recursive delete goes through the folder's contents and deletes the contents.  It calls itself recursively if it is a directory. Otherwise it deletes the file.

I removed the code that does not compile so that I can continue development on what was a working product.  Don't push code that does not even compile.